Elevator struct init, random floor number generated
So, I have been able to define the elevator struct as well as its fields that it will posses. I was also able to generate random floor values between 2-8 for each 10 passengers. The challenge I am facing and unsure to deal with is utilizing a thread to have the elevator move and check it's floor and the people who are supposed to exit there.